+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 - Issue #1527: statisticDashboard - restored WFS support for statistic dashboard and added default color palettes.
 - WFS-cluster: Fix a recursive style loop that triggered a ‘Maximum call stack size exceeded’ error.
 - Issue #1168: Fix autoRefresh for LayerGroups.
+- MapMarker: Fix marker visibility after ShareView.
 
 ---